requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
google-cloud-aiplatform==1.38.0
google-cloud-firestore==2.13.1
youtube-transcript-api==0.6.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
